fix(button/dropdown): fix breaking change caused by nesting Button inside of skeleton component
Previously, the skeleton wrapped the button (as it is intended to be used) but that changed the way
Button needed to be styled, which created a breaking change. With this change, the Button IS a
skeleton container, and includes the shimmer inside of it. So it uses the subcomponents of the
skeleton without directly using the react code.

fixes #308