feat(table): improve sorting typing, add third click to unsort, remove mutation of data prop
Add the SortState type, which indicates what column is sorted in in which direction. This replaces
the old, poorly-typed and documented two-element array (which is now deprecated, but is still
accepted). Also, this adds the SortDirection enum which improves code clarity and allows for the
third, unsorted state of a column. Finally, this adds the isSorted prop to get passed into each
header cell to allow for custom styling when a column is being sorted.

438